requirements.txt
langchain
langchain-openrouter
python-dotenv
deepagents
ddgs
langchain-community
playwright
langchain-core
langgraph
chainlit
langgraph-checkpoint-redis
beautifulsoup4